Changed Pause component flow, now exception & reload first
The previous flow didn't take into account whether adding an exception would help with the problem or not and did suggest that adding & reloading an exception would always fix the problem which is not always the case.

The new flow changes the order, exception adding and reloading takes place first, following feedback from the user the exception is now removed again in case it didn't fix the problem.